+++ src/main/java/gift/controller/ViewController.java
@@ -1,5 +1,7 @@
-package gift;
+package gift.controller;
 
+import gift.service.ProductService;
+import gift.model.Product;
 import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
