+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,5 +1,7 @@
-package gift;
+package gift.repository;
 
+import gift.exception.ProductNotFoundException;
+import gift.model.Product;
 import java.util.List;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
