+++ src/main/java/gift/service/ProductService.java
@@ -1,5 +1,7 @@
-package gift;
+package gift.service;
 
+import gift.model.Product;
+import gift.repository.ProductRepository;
 import java.util.List;
 import org.springframework.stereotype.Service;
