+++ src/main/java/gift/validation/RestrictedKeywordValidator.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.validation;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
