+++ src/test/java/gift/service/ProductServiceTest.java
@@ -1,8 +1,10 @@
-package gift;
+package gift.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import gift.exception.ProductNotFoundException;
+import gift.model.Product;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
